% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubg-client.R
\name{PUBGClient}
\alias{PUBGClient}
\title{PUBG API Client}
\description{
R6 class for interacting with the PUBG API
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiKey}}{PUBG API key}

\item{\code{baseUrl}}{Base URL for PUBG API}

\item{\code{platform}}{Gaming platform (default: "steam")}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PUBGClient-new}{\code{PUBGClient$new()}}
\item \href{#method-PUBGClient-getPlayerInfo}{\code{PUBGClient$getPlayerInfo()}}
\item \href{#method-PUBGClient-clone}{\code{PUBGClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PUBGClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-PUBGClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize PUBG client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PUBGClient$new(api_key, platform = "steam")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{Your PUBG API key}

\item{\code{platform}}{Gaming platform (default: "steam")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PUBGClient-getPlayerInfo"></a>}}
\if{latex}{\out{\hypertarget{method-PUBGClient-getPlayerInfo}{}}}
\subsection{Method \code{getPlayerInfo()}}{
Get player information for one or more players
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PUBGClient$getPlayerInfo(playerNames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{playerNames}}{Either a single player name or a vector of player names.
More than 10 players will be chunked into groups of 10 to avoid rate limiting.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing player information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PUBGClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PUBGClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PUBGClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
